refactor(football): split TransfermarktParser into parse_helpers/row_parser/parser
Split 393-line parser.py into inheritance chain:
- parse_helpers.py (120 lines): regex constants + extraction helpers
- row_parser.py (149 lines): _parse_row + profile parsing
- parser.py (136 lines): page-level facade methods
